Unify routes and menu items Limit rerenders of Routes, Menu, and updates to Login by memoization Generate routes dynamically with fallback to /login Disable links in menu from being clickable if disabled All disabled links route to '/'
