+++ client/src/components/Routers.js
@@ -1,55 +1,52 @@
 import React from 'react'
-import { Route, Switch } from 'react-router-dom'
+import { Redirect, Route, Switch } from 'react-router-dom'
 import Grid from '@material-ui/core/Grid'
-import TreeImageScrubber from './TreeImageScrubber'
-import Planters from './Planters'
-import Trees from './Trees'
 import Login from './Login'
-import Account from './Account'
-import Home from './Home'
-import Users from './Users'
-import SpeciesMgt from './SpeciesMgt'
-import { AppProvider, AppContext } from './Context'
+import { AppContext } from './Context'
 import PrivateRoute from './PrivateRoute'
-import TreeTest from './TreeTest'
 
 export default function Routers() {
   const refContainer = React.useRef()
+  const appContext = React.useContext(AppContext)
 
-  return (
-    <Grid container wrap="nowrap">
-      <Grid
-        item
-        style={{
-          width: '100%',
-        }}
-      >
+  return React.useMemo(() => {
+    return (
+      <Grid container wrap="nowrap">
         <Grid
-          container
-          ref={refContainer}
+          item
           style={{
-            height: '100vh',
-            overflow: 'auto',
+            width: '100%',
           }}
         >
-          <Switch>
-            <Route path="/login" component={Login} />
-            <PrivateRoute exact path="/" component={Home} />
-            <PrivateRoute path="/trees" component={Trees} />
-            <PrivateRoute path="/usermanager" component={Users} />
-            {/* <PrivateRoute path="/treetest" component={TreeTest} />Remove TreeTest if needed  */}
-            <PrivateRoute path="/trees" component={Trees} />
-            <PrivateRoute
-              path="/verify"
-              component={TreeImageScrubber}
-              getScrollContainerRef={() => refContainer.current}
-            />
-            <PrivateRoute path="/species" component={SpeciesMgt} />
-            <PrivateRoute path="/planters" component={Planters} />
-            <PrivateRoute path="/account" component={Account} />
-          </Switch>
+          <Grid
+            container
+            ref={refContainer}
+            style={{
+              height: '100vh',
+              overflow: 'auto',
+            }}
+          >
+            <Switch>
+              <Route path="/login" component={Login} />
+              { appContext.routes
+                .filter(({disabled}) => !disabled)
+                .map(({linkTo, exact = false, component}, idx) => (
+                  <PrivateRoute
+                    path={linkTo}
+                    component={component}
+                    exact={exact}
+                    getScrollContainerRef={linkTo === '/verify' ? () => refContainer.current : undefined}
+                    key={`route_${idx}`}
+                    />
+                ))
+              }
+              <Route>
+                <Redirect to="/login" />
+              </Route>
+            </Switch>
+          </Grid>
         </Grid>
       </Grid>
-    </Grid>
-  )
+    )
+  }, [appContext.routes])
 }
