+++ client/src/components/common/Menu.js
@@ -2,36 +2,16 @@ import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import Drawer from '@material-ui/core/Drawer'
 import MenuItem from '@material-ui/core/MenuItem'
-import IconSettings from '@material-ui/icons/Settings'
-import IconShowChart from '@material-ui/icons/ShowChart'
-import IconThumbsUpDown from '@material-ui/icons/ThumbsUpDown'
-import IconNature from '@material-ui/icons/Nature'
-import IconGroup from '@material-ui/icons/Group'
-import IconCompareArrows from '@material-ui/icons/CompareArrows'
-import IconPermIdentity from '@material-ui/icons/PermIdentity'
-import CategoryIcon from '@material-ui/icons/Category'
 import ListItemIcon from '@material-ui/core/ListItemIcon'
-import HomeIcon from '@material-ui/icons/Home'
 import ListItemText from '@material-ui/core/ListItemText'
 import Box from '@material-ui/core/Box'
 import { useTheme } from '@material-ui/styles'
 import IconLogo from '../IconLogo'
 import { AppContext } from '../Context'
-import { PERMISSIONS, hasPermission } from '../../models/auth'
 import { Link } from 'react-router-dom'
 import Grid from '@material-ui/core/Grid'
 
 export const MENU_WIDTH = 232
-
-const POLICIES = {
-  SUPER_PERMISSION: 'super_permission',
-  LIST_USER: 'list_user',
-  MANAGER_USER: 'manager_user',
-  LIST_TREE: 'list_tree',
-  APPROVE_TREE: 'approve_tree',
-  LIST_PLANTER: 'list_planter',
-  MANAGE_PLANTER: 'manage_planter',
-};
 
 const useStyles = makeStyles((theme) => ({
   drawer: {},
@@ -83,102 +63,19 @@ export default function GSMenu(props) {
   const classes = useStyles()
   const theme = useTheme()
   const appContext = React.useContext(AppContext)
-  const { user } = appContext
-  const menus = [
-    {
-      name: 'Home',
-      linkTo: '/',
-      icon: HomeIcon,
-      disabled: false,
-    },
-    {
-      name: 'Monitor',
-      linkTo: '/',
-      icon: IconShowChart,
-      disabled: true,
-    },
-    {
-      name: 'Verify',
-      linkTo: 'verify',
-      icon: IconThumbsUpDown,
-      disabled: !hasPermission(
-        user, [
-          POLICIES.SUPER_PERMISSION,
-          POLICIES.LIST_TREE,
-          POLICIES.APPROVE_TREE,
-        ]),
-    },
-    {
-      name: 'Trees',
-      linkTo: '/trees',
-      icon: IconNature,
-      disabled: !hasPermission(
-        user, [
-          POLICIES.SUPER_PERMISSION,
-          POLICIES.LIST_TREE,
 
-        ]),
-    },
-    {
-      name: 'Planters',
-      linkTo: 'planters',
-      icon: IconGroup,
-      disabled: !hasPermission(
-        user, [
-          POLICIES.SUPER_PERMISSION,
-          POLICIES.LIST_PLANTER,
-
-        ]),
-    },
-    {
-      name: 'Payments',
-      linkTo: '/',
-      icon: IconCompareArrows,
-      disabled: true,
-    },
-    {
-      name: 'Species',
-      linkTo: '/species',
-      icon: CategoryIcon,
-      //TODO this is temporarily, need to add species policy
-      disabled:
-        (!hasPermission(user, [
-          POLICIES.SUPER_PERMISSION,
-          POLICIES.LIST_TREE,
-        ])) || user.policy.organization !== undefined,
-    },
-    {
-      name: 'Settings',
-      linkTo: '/',
-      icon: IconSettings,
-      disabled: true,
-    },
-    {
-      name: 'User Manager',
-      linkTo: '/usermanager',
-      icon: IconGroup,
-      disabled: !hasPermission(
-        user, [
-          POLICIES.SUPER_PERMISSION,
-        ]),
-    },
-    {
-      name: 'Account',
-      linkTo: '/account',
-      icon: IconPermIdentity,
-      disabled: false,
-    },
-  ]
   const menu = (
     <>
       <Box p={4}>
         <IconLogo />
       </Box>
       <Box height={20} />
-      {menus.map((item, i) => (
-          <Link className={classes.linkItemText + (item.disabled ? ' ' + classes.disabledLinkItem : '')} to={`${item.disabled ? '/' : item.linkTo}`}>
+      {React.useMemo(() => (appContext.routes.map((item, i) => (
+          <Link
+            key={`menu_${i}`}
+            className={classes.linkItemText + (item.disabled ? ' ' + classes.disabledLinkItem : '')}
+            to={`${item.disabled ? '/' : item.linkTo}`}>
             <MenuItem
-              key={i}
               className={classes.menuItem}
               selected={props.active === item.name}
               disabled={item.disabled}
@@ -195,7 +92,7 @@ export default function GSMenu(props) {
                 </Grid>
             </MenuItem>
           </Link>
-      ))}
+      ))), [appContext.routes, props.active])}
     </>
   )
