+++ client/src/models/auth.js
@@ -4,8 +4,19 @@ const PERMISSIONS = {
   PLANTER: 3,
 };
 
+const POLICIES = {
+  SUPER_PERMISSION: 'super_permission',
+  LIST_USER: 'list_user',
+  MANAGER_USER: 'manager_user',
+  LIST_TREE: 'list_tree',
+  APPROVE_TREE: 'approve_tree',
+  LIST_PLANTER: 'list_planter',
+  MANAGE_PLANTER: 'manage_planter',
+};
+
 function hasPermission(user, p){
-  console.assert(user, "Why user fail?", user);
+  // Chris github - BirdTho 9/21 user may fail if called when logged out, as in the case in Context.js
+  // console.assert(user, "Why user fail?", user);
   if(!user) return false;
   if(p instanceof Array){
     return p.some(permission => {
@@ -22,4 +33,4 @@ function hasPermission(user, p){
 const session = {
 }
 
-export {PERMISSIONS, hasPermission, session};
+export {PERMISSIONS, POLICIES, hasPermission, session};
